requirements.txt
asgiref==3.6.0
Django==4.2
django-debug-toolbar==3.1.1
sqlparse==0.4.3
psycopg2-binary
djangorestframework
pillow
celery==4.4.7
django-celery-beat
python-dotenv==0.10.3